duration to ms
