+++ cmd/saveXcodeDerivedDataFiles.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/spf13/cobra"
 
 	xa "github.com/bitrise-io/bitrise-build-cache-cli/internal/analytics"
@@ -55,7 +54,7 @@ var saveXcodeDerivedDataFilesCmd = &cobra.Command{
 				errStr := cmdError.Error()
 				op.Error = &errStr
 			}
-			op.Duration = int(time.Since(op.StartedAt))
+			op.Duration = int(time.Since(op.StartedAt).Milliseconds())
 
 			xaClint, clientErr := xa.NewClient(consts.AnalyticsServiceEndpoint, authConfig.AuthToken, logger)
 			if clientErr != nil {
@@ -109,7 +108,7 @@ func saveXcodeDerivedDataFilesCmdFn(ctx context.Context, authConfig common.Cache
 	}
 	logger.Infof("(i) Cache key: %s", cacheKey)
 
-	op := newCacheOperation(startT, cacheKey, envProvider)
+	op := newCacheOperation(startT, xa.OperationTypeUpload, cacheKey, envProvider)
 
 	kvClient, err := createKVClient(ctx, authConfig, envProvider, logger)
 	if err != nil {
@@ -193,59 +192,4 @@ func saveXcodeDerivedDataFilesCmdFn(ctx context.Context, authConfig common.Cache
 	}
 
 	return op, nil
-}
-
-func newCacheOperation(startT time.Time, cacheKey string, envProvider func(string) string) *xa.CacheOperation {
-	op := &xa.CacheOperation{
-		OperationID:   uuid.NewString(),
-		OperationType: xa.OperationTypeUpload,
-		StartedAt:     startT,
-		CacheKey:      cacheKey,
-		CIProvider:    "bitrise",
-		CLIVersion:    envProvider("BITRISE_BUILD_CACHE_CLI_VERSION"),
-		CommitHash:    envProvider("BITRISE_GIT_COMMIT"),
-	}
-
-	projectID := envProvider("BITRISE_APP_SLUG")
-	if projectID != "" {
-		op.ProjectID = &projectID
-	}
-
-	buildID := envProvider("BITRISE_BUILD_SLUG")
-	if buildID != "" {
-		op.BuildID = &buildID
-	}
-
-	workflowID := envProvider("BITRISE_TRIGGERED_WORKFLOW_ID")
-	if workflowID != "" {
-		op.WorkflowID = &workflowID
-	}
-
-	workflowTitle := envProvider("BITRISE_TRIGGERED_WORKFLOW_TITLE")
-	if workflowTitle != "" {
-		op.WorkflowTitle = &workflowTitle
-	}
-
-	repoURL := envProvider("GIT_REPOSITORY_URL")
-	if repoURL != "" {
-		op.RepositoryURL = &repoURL
-	}
-
-	branch := envProvider("BITRISE_GIT_BRANCH")
-	if branch != "" {
-		op.Branch = &branch
-	}
-
-	return op
-}
-
-func fillCacheOperationWithUploadStats(op *xa.CacheOperation, stats xcode.UploadFilesStats) {
-	op.TransferSize = stats.UploadSize
-	op.FileStats = xa.FileStats{
-		FilesToTransfer:  stats.FilesToUpload,
-		FilesTransferred: stats.FilesUploaded,
-		FilesFailed:      stats.FilesFailedToUpload,
-		FilesMissing:     stats.FilesToUpload,
-		TotalFiles:       stats.TotalFiles,
-	}
 }
